fix: EarlyStoppingCallback 클래스 역모듈화
#37
